refactor(admin): address structural audit findings
- Re-export schemas and types from declarations barrel
- Replace useForm with useZodForm in SearchForm
- Extract formatDate, formatDateTime, STATUS_LABELS to shared/constants
- Extract Pagination component to shared/Pagination
- Extract detail sections to DetailSections.tsx
- Add DeleteModal unit tests
