+++ packages/app/src/modules/admin/declarations/index.ts
@@ -1,2 +1,11 @@
 export { AdminDeclarationDetailPage } from "./AdminDeclarationDetailPage";
 export { AdminDeclarationsPage } from "./AdminDeclarationsPage";
+export type {
+	SearchDeclarationsInput,
+	SearchDeclarationsOutput,
+} from "./schemas";
+export {
+	deleteDeclarationsSchema,
+	getDeclarationByIdSchema,
+	searchDeclarationsSchema,
+} from "./schemas";
